requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
selenium
python-dateutil
webdriver-manager
